feat(ui-sheets): enhance `ContextMenuInput`
