+++ packages/design/src/components/input-number/index.module.less
@@ -6,6 +6,7 @@
     font-size: 12px;
     line-height: 26px;
     vertical-align: middle;
+    background-color: rgb(var(--bg-color-secondary));
     border: 1px solid rgb(var(--border-color));
     border-radius: var(--border-radius-base);
     transition: all 0.3s;
@@ -116,7 +117,6 @@
 
     &-disabled {
         .input-number-input {
-            background-color: rgb(var(--bg-color-secondary));
             cursor: not-allowed;
         }
         .input-number-handler {
