+++ packages/ui-plugin-sheets/src/View/ContextMenu/index.module.less
@@ -1,3 +1,10 @@
+.ui-plugin-sheets-context-menu-input {
+    display: flex;
+    align-items: center;
+    gap: var(--margin-sm);
+}
+
+// TODO: may be conflict with other styles, should be more specific
 .context-menu {
     position: fixed;
